add image and log out to menu
