+++ ui/components/global/AuthBadge/AuthBadge.module.scss
@@ -11,37 +11,21 @@
     width: calc(100% - 46px);
   }
   &__ava {
-    box-sizing: border-box;
+    display: flex;
+    align-items: center;
+    justify-content: center;
     width: 34px;
     height: 34px;
     margin-left: 12px;
     background: var(--gray);
     border-radius: 100%;
     overflow: hidden;
-    position: relative;
-    &:before,
-    &:after {
-      content: "";
-      border-radius: 100%;
-      margin: auto;
-      position: absolute;
-      left: 0;
-      right: 0;
-      top: 0;
-      bottom: 0;
-      background: var(--white-20);
-      width: 14px;
-      height: 14px;
-    }
-    &:after {
-      transform: scale(1.5) translateY(83%);
-    }
+
     img {
-      max-width: 100%;
-      height: auto;
       border-radius: 100%;
-      position: relative;
-      z-index: 2;
+      width: 100%;
+      height: auto;
+      background-size: contain;
     }
   }
   &:hover {
