+++ ui/components/global/AuthBadge/index.jsx
@@ -2,8 +2,10 @@ import Link from 'next/link';
 import Image from 'next/image';
 import { useSession } from 'next-auth/react';
 
-import styles from './AuthBadge.module.scss';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
+import { faUser } from '@fortawesome/free-solid-svg-icons';
+import styles from './AuthBadge.module.scss';
 const AuthBadge = () => {
   const { data: session } = useSession();
 
@@ -24,15 +26,10 @@ const AuthBadge = () => {
           : 'Авторизация'}
       </div>
       <div className={styles.authBadge__ava}>
-        {session?.user?.image && (
-          <Image
-            src='profilePic.webp'
-            width={34}
-            height={34}
-            alt='User profile picture'
-            quality={80}
-            loader={customLoader}
-          />
+        {session?.user?.image ? (
+          <img src={process.env.NEXT_PUBLIC_API_URL + session?.user?.image} />
+        ) : (
+          <FontAwesomeIcon icon={faUser} />
         )}
       </div>
     </Link>
